pyproject.toml
[tool.poetry]
name = "pycodebuddy"
version = "0.0.1"
description = "A command-line tool leveraging OpenAI's GPT for intelligent project management within a Git-integrated shell."
authors = ["Cem Basoglu <cem.basoglu@web.de>"]
license = "MIT"
readme = "README.md"
keywords = ["GPT", "automation", "OpenAI", "interactive shell", "Git", "project management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = ">=3.9"
openai = ">=1.54.4"
pyyaml = ">=6.0.2"
gitignore-parser = "^0.1.11"

[tool.poetry.scripts]
codebuddy = "pycodebuddy.main:main"
cb = "pycodebuddy.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
